:art: Renamed the Assets module.
